requirements.txt
streamlit
google-genai
faiss-cpu
vaderSentiment
wordcloud
plotly
pandas
numpy
nltk
scikit-learn